Move wallet actions to folder
