+++ ui/core/src/usecases/wallet/eth.test.ts
@@ -1,8 +1,8 @@
-import createActions from "./ethWallet";
+import createActions from "./eth";
 
-import { Address, Asset, Network, TxParams } from "../entities";
+import { Address, Asset, Network, TxParams } from "../../entities";
 import { Msg } from "@cosmjs/launchpad";
-import { IWalletService } from "../services/IWalletService";
+import { IWalletService } from "../../services/IWalletService";
 let mockEthereumService: IWalletService & {};
 let mockEventBusService: any;
 let ethWalletActions: ReturnType<typeof createActions>;
