+++ ui/core/src/usecases/wallet/sif.ts
@@ -1,7 +1,7 @@
-import { Address, TxParams } from "../entities";
+import { Address, TxParams } from "../../entities";
 import { validateMnemonic } from "bip39";
-import { Mnemonic } from "../entities/Wallet";
-import { UsecaseContext } from ".";
+import { Mnemonic } from "../../entities/Wallet";
+import { UsecaseContext } from "..";
 import { effect } from "@vue/reactivity";
 
 export default ({
